+++ AffiseAttributionLib/Classes/internal/BuildConfig.swift
@@ -1,5 +1,5 @@
 import Foundation
 
 internal struct BuildConfig {
-    static let AFFISE_VERSION = "1.6.42"
+    static let AFFISE_VERSION = "1.6.43"
 }