+++ AffiseAttributionLib/Classes/usecase/IndexUseCase.swift
@@ -1,0 +1,4 @@
+protocol IndexUseCase {
+    func getUuidIndex() -> Int64
+    func getAffiseEventIdIndex() -> Int64
+}