+++ AffiseAttributionLib/Classes/usecase/SendDataToServerUseCase.swift
@@ -3,5 +3,5 @@ protocol SendDataToServerUseCase {
     /**
      * Send
      */
-    func send(withDelay: Bool)
+    func send(withDelay: Bool, sendEmpty: Bool)
 }